refactor: split dashboard into tabbed components
